api/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic>=2.0.0,<3.0.0
pydantic-settings>=2.0.0
sqlalchemy==2.0.28
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
alembic==1.13.1
psycopg2-binary==2.9.9
pymongo>=4.5.0
weaviate-client>=3.26.7,<4.0.0
python-dotenv==1.0.1
pytest==8.0.2
httpx==0.27.0
email-validator==2.2.0 